% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muhat_twfe.R
\name{muhat_twfe}
\alias{muhat_twfe}
\title{Estimate \eqn{\hat{\mu}_{it}(1)} and \eqn{\hat{\mu}_{it}(0)} from a TWFE regression with cross-fitting}
\usage{
muhat_twfe(
  Y,
  tr,
  X_tv = NULL,
  X_ti = NULL,
  main_tvc = NULL,
  main_tic = NULL,
  int_tvc = NULL,
  int_tic = NULL,
  joint = TRUE,
  cf = TRUE,
  nfolds = 10,
  foldid = NULL
)
}
\arguments{
\item{Y}{an nxT matrix for the outcome variable}

\item{tr}{an nxT binary matrix for the treatment variable (which must be staggered)}

\item{X_tv}{a length-T list of nxp matrices where p is the number of time-varying variables. The t-th element is the matrix \eqn{(X_{itj})_{i\in [n], j\in [p]}}}

\item{X_ti}{an nxq matrix where q is the number of time-invariant variables}

\item{main_tvc}{a vector of variable names for (both time-varying and time-invariant) variables with time-varying coefficients in the non-interacted terms. Leave it as default if no time-varying coefficient is used.}

\item{main_tic}{a vector of variable names for (both time-varying and time-invariant) variables with time-invariant coefficients in the non-interacted terms. Leave it as default if no time-invariant coefficient is used.}

\item{int_tvc}{a vector of variable names for (both time-varying and time-invariant) variables with time-varying coefficients in the interacted terms. Leave it as default if no time-varying coefficient is used.}

\item{int_tic}{a vector of variable names for (both time-varying and time-invariant) variables with time-invariant coefficients in the interacted terms. Leave it as default if no time-invariant coefficient is used.}

\item{joint}{TRUE if treatment and control groups are jointly fit. See Details.}

\item{cf}{TRUE iff cross-fitting is used}

\item{nfolds}{number of folds (K in Algorithm 1)}

\item{foldid}{NULL if fold indices are generated randomly or a list of length K with the k-th element a vector of indices in fold k}
}
\value{
a list of form list(muhat1 = , muhat0 = ) where muhat1 and muhat0 are both nxT matrices
}
\description{
\code{muhat_twfe} fits a TWFE regression time-varying/invariant variables and time-varying/invariant coefficients in both non-interacted and interacted terms.
To formally write down the model, we define the following quantities
\itemize{
\item{\eqn{X_{itj}}}{ j-th time-varying covariate \eqn{(j=1,\ldots,p)} for unit i at time t \eqn{(j=p+1,\ldots, p+q)}}
\item{\eqn{X_{ik}}}{ j-th time-invariant covariate for unit i \eqn{(j=p+1,\ldots, p+q)}}
\item{\eqn{S_{main\_tvc}\subset \{1, \ldots, p+q\}}}{ the subset of variables with time-varying coefficients in non-interacted terms}
\item{\eqn{S_{main\_tic}\subset \{1, \ldots, p+q\}}}{ the subset of variables with time-invariant coefficients in non-interacted terms}
\item{\eqn{S_{int\_tvc}\subset \{1, \ldots, p+q\}}}{ the subset of variables with time-varying coefficients in interacted terms}
\item{\eqn{S_{int\_tic}\subset \{1, \ldots, p+q\}}}{ the subset of variables with time-invariant coefficients in interacted terms}
}
When \code{joint=TRUE}, we fit the following TWFE regression on the treated and control groups together:
\deqn{Y_{it} = \alpha_i + \lambda_t + \sum_{j\in S_{main\_tvc}, j\le p}X_{itj}\beta_{jt} + \sum_{j\in S_{main\_tvc}, j > p}X_{ij}\beta_{jt}
+ \sum_{j\in S_{main\_tic}, j\le p}X_{itj}\beta_{j} + \sum_{j\in S_{main\_tic}, j > p}X_{ij}\beta_{j} + \tau W_{it} + \sum_{j\in S_{int\_tvc}, j\le p}W_{it}X_{itj}\phi_{jt}
+ \sum_{j\in S_{int\_tvc}, j > p}W_{it}X_{ij}\phi_{jt} + \sum_{j\in S_{int\_tic}, j\le p}W_{it}X_{itj}\phi_{j} + \sum_{j\in S_{int\_tic}, j > p}W_{it}X_{ij}\phi_{j} + \epsilon_{it}.}
With this fit,
\deqn{\hat{\mu}_{it}(0) =  \sum_{j\in S_{main\_tvc}, j\le p}X_{itj}\beta_{jt} + \sum_{j\in S_{main\_tvc}, j > p}X_{ij}\beta_{jt}
+ \sum_{j\in S_{main\_tic}, j\le p}X_{itj}\beta_{j} + \sum_{j\in S_{main\_tic}, j > p}X_{ij}\beta_{j},}
and
\deqn{\hat{\mu}_{it}(1) = \hat{\mu}_{it}(0) + \sum_{j\in S_{int\_tvc}, j\le p}X_{itj}\phi_{jt}
+ \sum_{j\in S_{int\_tvc}, j > p}X_{ij}\phi_{jt} + \sum_{j\in S_{int\_tic}, j\le p}X_{itj}\phi_{j} + \sum_{j\in S_{int\_tic}, j > p}X_{ij}\phi_{j}.}
When \code{joint=FALSE}, we fit the following TWFE regression for the treated and control groups separately:
\deqn{Y_{it} = \alpha_i + \lambda_t + \sum_{j\in S_{main\_tvc}, j\le p}X_{itj}\beta_{jt} + \sum_{j\in S_{main\_tvc}, j > p}X_{ij}\beta_{jt}
+ \sum_{j\in S_{main\_tic}, j\le p}X_{itj}\beta_{j} + \sum_{j\in S_{main\_tic}, j > p}X_{ij}\beta_{j} + \epsilon_{it}.}
In particular, \code{main_tic} and \code{int_inc} will not be used.
}
\examples{
\donttest{
# Load opentable data
data(opentable)
library("tidyverse")

# Generate the treatment assignments
Y <- opentable \%>\% select(state, time, reserv.diff) \%>\%
 spread(time, reserv.diff) \%>\%
 select(-state) \%>\%
 as.matrix
tr <- opentable \%>\% select(state, time, treat) \%>\%
 spread(time, treat) \%>\%
 select(-state) \%>\%
 as.matrix
T <- ncol(tr)

## Construct covariate matrices
# vote, beds, and region are time-invariant covariates
X_vote <- opentable$vote \%>\%
  matrix(nrow = T) \%>\% t \%>\% .[, 1]
X_beds <- opentable$beds \%>\%
  matrix(nrow = T) \%>\% t \%>\% .[, 1]
X_region <- opentable$region \%>\%
  matrix(nrow = T) \%>\% t \%>\% .[, 1] \%>\%
  data.frame(region = .) \%>\%
  model.matrix(~0+region, data = .) \%>\%
  .[, -1] \%>\% as.matrix
X_ti <- cbind(vote = X_vote, beds = X_beds, X_region)

# confirmed is a time-varying covariate
X_confirmed <- opentable$confirmed \%>\%
  matrix(nrow = T) \%>\% t
X_tv <- lapply(1:T, function(t){
  tmp <- as.matrix(X_confirmed[, t])
  colnames(tmp) <- "confirmed"
  tmp
})

# Using cross-fitting and randomly generated folds
set.seed(2023)
nfolds <- 10
main_tic <- "confirmed"
int_tic <- c("confirmed", colnames(X_ti))
muhat <- muhat_twfe(Y, tr, X_tv, X_ti, main_tic = main_tic, int_tic = int_tic,
                    joint = TRUE, cf = TRUE, nfolds = nfolds)

# Using cross-fitting and randomly generated folds
set.seed(2023)
foldid <- ripw:::gen_cf_inds(nrow(Y), nfolds)
muhat2 <- muhat_twfe(Y, tr, X_tv, X_ti, main_tic = main_tic, int_tic = int_tic,
                     joint = TRUE, foldid = foldid)
identical(muhat, muhat2)

# Not using cross-fitting
muhat3 <- muhat_twfe(Y, tr, X_tv, X_ti, main_tic = main_tic, int_tic = int_tic,
                     joint = TRUE, cf = FALSE)
}
}
