% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opentable}
\alias{opentable}
\title{Opentable dataset}
\format{
An object of class \code{data.frame} with 504 rows and 9 columns.
}
\usage{
opentable
}
\description{
The dataset analyzed in Section 5.2. It includes the following variables.
\itemize{
\item state
\item date
\item time: number of days since 02/29/2020
\item reserv.diff: reservation difference
\item confirmed: number of confirmed cases
\item treat: whether state of emergency has been declared
\item vote: vote share in percentage points
\item beds: log(number of hospital beds)
\item region: region of the state (South, West, North Central, and Northeast)
}
}
\keyword{datasets}
