% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_equation.R
\name{solve_date_maxmin}
\alias{solve_date_maxmin}
\title{Solver for the DATE equation that maximizes min_{w}Pi(w) (Appendix C.5)}
\usage{
solve_date_maxmin(
  support,
  xi,
  init = rep(1, ncol(support))/ncol(support),
  tol = 1e-06
)
}
\arguments{
\item{support}{the matrix \eqn{(\tilde{w}_{(1)}, ..., \tilde{w}_{(K)})}}

\item{xi}{weights used for the DATE}

\item{init}{initial value}

\item{tol}{tolerance for the objective value}
}
\value{
\item{sol}{a solution (if found)}
\item{val}{the value of the objective at sol}
\item{init_val}{the value of the objective at init}
}
\description{
Solver for the DATE equation that maximizes min_{w}Pi(w) (Appendix C.5)
}
