% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_equation.R
\name{solve_date_maxmin}
\alias{solve_date_maxmin}
\title{Solver for the DATE equation that maximizes \eqn{min_{w}\Pi(w)} (Appendix C.5)}
\usage{
solve_date_maxmin(
  support,
  xi,
  init = rep(1, ncol(support))/ncol(support),
  tol = 1e-06
)
}
\arguments{
\item{support}{the TxK matrix \eqn{(\tilde{w}_{(1)}, ..., \tilde{w}_{(K)})}}

\item{xi}{weights used for the DATE}

\item{init}{initial value}

\item{tol}{tolerance for the objective value}
}
\value{
\item{sol}{a solution (if found)}
\item{val}{the value of the objective at sol}
\item{init_val}{the value of the objective at init}
}
\description{
Solver for the DATE equation that maximizes \eqn{min_{w}\Pi(w)} (Appendix C.5)
}
\examples{
\donttest{
# Generate the support S* for staggered treatment with T=4
support <- cbind(rep(0, 4), c(rep(0, 3), 1),
                 c(rep(0, 2), rep(1, 2)),
                 c(0, rep(1, 3)), rep(1, 4))
xi <- rep(1/4, 4)
solve_date_maxmin(support, xi)

# Generate the support S* for transient treatment with T=4
support <- cbind(rep(0, 4), c(0, 0, 0, 1),
                 c(0, 0, 1, 0), c(0, 1, 0, 0),
                 c(1, 0, 0, 0))
xi <- rep(1/4, 4)
solve_date_maxmin(support, xi)
}

}
